requirements.txt
Flask
requests
openai
pandas
python-dotenv
